Test the chat event listener's transaction phase without a database context
